+++ backend/src/main.ts
@@ -1,27 +1,30 @@
-import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
-import { ValidationPipe } from '@nestjs/common';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { NestFactory } from "@nestjs/core";
+import { AppModule } from "./app.module";
+import { ValidationPipe } from "@nestjs/common";
+import { DocumentBuilder, SwaggerModule } from "@nestjs/swagger";
 
 async function bootstrap() {
-    const PORT = process.env.PORT || 8888;
+  const PORT = process.env.PORT || 8888;
 
-    const app = await NestFactory.create(AppModule);
+  const app = await NestFactory.create(AppModule);
+  app.enableCors({
+    origin: true,
+    methods: "GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS",
+    credentials: true,
+  });
 
-    const config = new DocumentBuilder()
-        .setTitle('Social network')
-        .setDescription('REST API documentation')
-        .setVersion('1.0.0')
-        .addTag('Oleh')
-        .build();
+  const config = new DocumentBuilder()
+    .setTitle("Social network")
+    .setDescription("REST API documentation")
+    .setVersion("1.0.0")
+    .addTag("Oleh")
+    .build();
 
-    const document = SwaggerModule.createDocument(app, config);
-    SwaggerModule.setup('/api/docs', app, document);
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup("/api/docs", app, document);
 
-    app.useGlobalPipes(new ValidationPipe());
+  app.useGlobalPipes(new ValidationPipe());
 
-    await app.listen(PORT, () =>
-        console.log(`Server started on port = ${PORT}`),
-    );
+  await app.listen(PORT, () => console.log(`Server started on port = ${PORT}`));
 }
 bootstrap();