+++ frontend/src/pages/Login/SignIn.tsx
@@ -11,14 +11,19 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import styles from './Login.module.css';
 import { validationSchema } from '../../utils/validators/signIn';
 import { Button } from '@mui/material';
+import { useDispatch } from "react-redux";
+import { fetchedAuth } from "../../redux/ducks/auth/auth.slice";
 
 interface SignInInterface {
   onModelChange: () => void;
 }
 
 export const SignIn: React.FC<SignInInterface> = ({ onModelChange }) => {
+
+  const dispatch = useDispatch();
   const onSubmit = (event: any) => {
-    return null;
+    debugger;
+    dispatch(fetchedAuth(event))
   };
   const {
     register, control, handleSubmit, formState: { errors },