+++ frontend/src/redux/ducks/auth/action.types.ts
@@ -2,25 +2,24 @@ import { Action } from "@reduxjs/toolkit";
 import { LoginDto } from "../../../types/auth/login.dto";
 
 
-export enum HistoricalDataActionsTypes {
-    REQUEST_AUTH = 'REQUEST_AUTH',
+export enum AuthActionTypes {
     AUTH_SUCCESS = 'AUTH_SUCCESS',
     AUTH_FAILED = 'AUTH_FAILED',
     FETCHED_AUTH = 'FETCHED_AUTH',
 }
 
 
-export interface IAuthSuccess extends Action<HistoricalDataActionsTypes>{
-    type: HistoricalDataActionsTypes.AUTH_SUCCESS,
+export interface IAuthSuccess extends Action<AuthActionTypes>{
+    type: AuthActionTypes.AUTH_SUCCESS,
     token: string
 }
-export interface IAuthFailed extends Action<HistoricalDataActionsTypes>{
-    type: HistoricalDataActionsTypes.AUTH_FAILED,
+export interface IAuthFailed extends Action<AuthActionTypes>{
+    type: AuthActionTypes.AUTH_FAILED,
     error: any
 }
-export interface IFetchedAuth extends Action<HistoricalDataActionsTypes>{
-    type: HistoricalDataActionsTypes.FETCHED_AUTH,
-    authData: LoginDto
+export interface IFetchedAuth extends Action<AuthActionTypes>{
+    type: AuthActionTypes.FETCHED_AUTH,
+    payload: LoginDto
 }